requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
neo4j==5.23.1
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
